Casting error fix

+++ src/Game/GameMacros.hpp
@@ -23,5 +23,5 @@ const float _player_acceleration = 0.1f;
 const float _player_max_velocity = 4.f;
 const float _player_movement_mult = -1.f;
 const float _player_movement_trans = 21.f;
-const float _pi_fourth = sqrt(2)/2.f;
+const float _pi_fourth = static_cast<float>(sqrt(2))/2.f;
 #endif
